% Please edit documentation in R/shiny.R
\name{appendCheckboxes}
\alias{appendCheckboxes}
\alias{checkboxRows}
\title{Append a column of checkboxes to the data for selecting rows}
\usage{
appendCheckboxes(data, after = TRUE)

checkboxRows(data)
}
\arguments{
\item{data}{the data to be displayed as the table}

\item{after}{add the checkboxes after the last column (\code{TRUE}) or before
the first column (\code{FALSE}) of the table}
}
\description{
Append a column of checkboxes with the CSS class \samp{DT checkboxRows} to
the data. The checkboxes can be used for selecting rows in Shiny. Each
checkbox has a data field \samp{data-row} that contains the index of the
current row. The HTML code for checkboxes are generated by
\code{checkboxRows()}.
}
\details{
For a table output with id \code{foo} in Shiny, the indices of the selected
rows are stored in \code{input$foo_selected}.
}
\examples{
library(DT)
datatable(appendCheckboxes(iris), escape = -7)
}
\references{
See \url{http://rstudio.github.io/DT/shiny.html} for examples.
}
