% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b-spline-basic-radial-basis.R
\name{b_spline_basic_radial_basis}
\alias{b_spline_basic_radial_basis}
\title{A basic radial b-spline basis}
\usage{
b_spline_basic_radial_basis(x, k, h = 1, s = 1)
}
\arguments{
\item{x}{location to evaluate at}

\item{k}{location of basis knot}

\item{h}{height parameter}

\item{s}{knot width/scaling parameter}
}
\value{
value at x for knot at k with parameter h
}
\description{
A basic radial b-spline basis
}
