% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b-spline-covariate-repr.R
\name{b_spline_covariate_repr}
\alias{b_spline_covariate_repr}
\title{Represent the matrix of covariates for the b-spline}
\usage{
b_spline_covariate_repr(
  x,
  knots,
  basis = dt:::b_spline_basic_radial_basis,
  label = "spline",
  ...
)
}
\arguments{
\item{x}{values to calculate spline values for}

\item{knots}{knot locations}

\item{basis}{basis function to use for calculating the spline values}

\item{label}{1-length character vector used to prefix the resulting
set of column names}

\item{...}{further arguments to the basis function}
}
\value{
a representation of the spline value per x value
}
\description{
Represent the matrix of covariates for the b-spline
}
