% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b-spline-integrated-basic-radial-basis.R
\name{b_spline_integrated_basic_radial_basis}
\alias{b_spline_integrated_basic_radial_basis}
\title{Integral of a basic radial b-spline basis}
\usage{
b_spline_integrated_basic_radial_basis(lb, ub, k, h = 1)
}
\arguments{
\item{lb}{lower bound of integration}

\item{ub}{upper bound of integration}

\item{k}{location of basis knot}

\item{h}{height parameter}
}
\value{
integral between lb and ub for knot at k with parameter h
}
\description{
Integral of a basic radial b-spline basis
}
