% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-flat-linear-model.R
\name{build_flat_linear_model}
\alias{build_flat_linear_model}
\title{(Lazily) build a model from its components}
\usage{
build_flat_linear_model(...)
}
\arguments{
\item{...}{any components for a single model}
}
\value{
(lazy) model representation
}
\description{
(Lazily) build a model from its components
}
