% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-flat-model-component.R
\name{build_flat_model_component}
\alias{build_flat_model_component}
\title{Construct a complete flat model component}
\usage{
build_flat_model_component(
  type,
  format = "row",
  N,
  K,
  n_nze,
  nze_idx,
  nze_value,
  row_start_idx,
  row_n_nze,
  row_nze_idx,
  col_nze_idx,
  row_names,
  col_names,
  label,
  ...
)
}
\arguments{
\item{type}{character string listing the component type}

\item{format}{character string listing the format ('row', always)}

\item{N}{single integer, specify the number of observations in the
component.}

\item{K}{single integer, specify the number of parameters in the model
component}

\item{n_nze, }{single integer, specify the number of non-zero elements in
the flat representation of the component model matrix}

\item{nze_idx, }{vector of integers, specify the indexes of non-zero entries
in the vector of non-zero values}

\item{nze_value, }{vector of non-zero values (floating point numbers)}

\item{row_start_idx, }{vector of integers, each indicates the start of a row
in the vector of non-zero values}

\item{row_n_nze, }{number of non-zero entries in each row}

\item{row_nze_idx, }{row of each entry in the vector of non-zero values}

\item{col_nze_idx, }{col of each entry in the vector of non-zero values}

\item{row_names, }{row names for the model component}

\item{col_names, }{column names for the model component}

\item{label, }{model component label}

\item{reference, }{reference category relevant for 'discrete-offset' type
only.}
}
\value{
flat model component representation
}
\description{
Enforces some of the list structure that's implicit currently
}
