% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbind-flat-matrices.R
\name{cbind_flat_matrices}
\alias{cbind_flat_matrices}
\title{Combine two matrices into a single (flat) matix}
\usage{
cbind_flat_matrices(a, ...)
}
\arguments{
\item{a}{first flat matrix}

\item{...}{additional flat matrices}

\item{b}{second flat matrix}
}
\value{
flat (row-wise) matrix, cbind(a,b)
}
\description{
Combine two matrices into a single (flat) matix
}
