% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbind-flat-vectors.R
\name{cbind_flat_vectors}
\alias{cbind_flat_vectors}
\title{Combine two flat vectors into a single (flat) matix}
\usage{
cbind_flat_vectors(a, b)
}
\arguments{
\item{a}{first column vector}

\item{b}{second column vector}
}
\value{
flat (row-wise) matrix, cbind(a,b)
}
\description{
Combine two flat vectors into a single (flat) matix
}
