% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{datatableThemeVariables}
\alias{datatableThemeVariables}
\alias{datatableThemeVariablesClear}
\title{Datatable theming variables}
\usage{
datatableThemeVariables(...)

datatableThemeVariablesClear()
}
\arguments{
\item{...}{Variables to be set, using \code{name = value}; or, a single unnamed
argument which is a named list of variables to set. To specify a name that
uses characters that aren't valid for R identifiers, wrap that name in
quotes, double quotes, or backticks.}
}
\value{
If no arguments are passed, a list of all set options is returned. If
arguments are passed, a list containing the previous values of the newly
set options is returned invisibly.
}
\description{
Use \code{datatableThemeVariables()} to set Datatable SASS variables to influence their
default CSS styling. Use \code{datatableThemeVariablesClear()} to clear any currently
set variables.
}
\examples{
library(htmltools)
# Color palette inspired by https://bootswatch.com/solar/
color <- "#002B36"
stroke <- "#839496"

# Full list of SASS variables
# https://github.com/DataTables/DataTablesSrc/blob/master/css/jquery.dataTables.scss
datatableThemeVariables(
  "table-header-border" = color,
  "table-body-border" = color,
  "table-row-background" = color,
  "table-control-color" = stroke,
  "table-paging-button-active" = stroke,
  "table-paging-button-hover" = color,
  "table-shade" = stroke
)
browsable(
  tags$body(
    style = sprintf("background-color: \%s; color: \%s", color, stroke),
    datatable(mtcars)
  )
)

}
