% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-intercept.R
\name{discrete_intercept}
\alias{discrete_intercept}
\title{Simple discrete intercept}
\usage{
discrete_intercept(x, label = "x")
}
\arguments{
\item{x}{value to use for covariate}

\item{label}{1-length character vector used as prefix for column names}
}
\value{
a matrix representing the covariate
}
\description{
Simple discrete intercept
}
