% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-intercept-repr.R
\name{discrete_intercept_repr}
\alias{discrete_intercept_repr}
\title{Represent a simple discrete intercept}
\usage{
discrete_intercept_repr(x, label = "x")
}
\arguments{
\item{x}{value to use for covariate}

\item{label}{1-length character vector used as prefix for column names}
}
\value{
a representation of the covariate
}
\description{
Represent a simple discrete intercept
}
