% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-offset.R
\name{discrete_offset}
\alias{discrete_offset}
\title{Represent a simple discrete offset}
\usage{
discrete_offset(x, reference, label = "x")
}
\arguments{
\item{x}{value to use for offset}

\item{reference}{value defined as non-effect}

\item{label}{1-length character vector used as prefix for column names}
}
\value{
a representation of the covariate
}
\description{
Represent a simple discrete offset
}
