\name{DT-imports}
\alias{JS}
\alias{saveWidget}
\alias{\%>\%}
\docType{import}
\title{Objects imported from other packages}
\description{
These objects are imported from other packages. Follow the links to their documentation.
\describe{
  \item{htmlwidgets}{\code{\link[htmlwidgets:JS]{JS}}, \code{\link[htmlwidgets:saveWidget]{saveWidget}}}
  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}
}}
