% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-package.R
\docType{package}
\name{dt-package}
\alias{dt}
\alias{dt-package}
\title{dt: Core modeling tools}
\description{
Contains tools for modeling basic components of various datasets with a focus on the time domain, hazard-based models, and flexible approaches.
}
\author{
\strong{Maintainer}: Krzysztof Sakrejda \email{krzysztof@mostek.io} (\href{https://orcid.org/0000-0001-7587-355X}{ORCID})

}
\keyword{internal}
