% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat-interact.R
\name{flat_interact}
\alias{flat_interact}
\title{Create an interaction term between two flat represented model components}
\usage{
flat_interact(a, b)
}
\arguments{
\item{a}{first model component}

\item{b}{second model component}
}
\value{
interaction resolved component
}
\description{
Create an interaction term between two flat represented model components
}
