% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat-matrix-times-column-vector.R
\name{flat_matrix_times_column_vector}
\alias{flat_matrix_times_column_vector}
\title{Multiple a flat matrix by a flat column vector (col-wise)}
\usage{
flat_matrix_times_column_vector(m, v)
}
\arguments{
\item{m}{flat matrix}

\item{v}{flat column vector}
}
\value{
column-wise product
}
\description{
Multiple a flat matrix by a flat column vector (col-wise)
}
