% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat-permute.R
\name{flat_permute}
\alias{flat_permute}
\title{Combine two flat linear model components by column-wise permutation
/ product}
\usage{
flat_permute(a, b)
}
\arguments{
\item{a, }{a first model components}

\item{b, }{a second model component}
}
\value{
permutation of submodel columns
}
\description{
e.g.-an interaction
}
