% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat-pick-matrix-column.R
\name{flat_pick_matrix_column}
\alias{flat_pick_matrix_column}
\title{Pick out a column from a flat matrix}
\usage{
flat_pick_matrix_column(m, j)
}
\arguments{
\item{m}{matrix}

\item{j}{index of column to pick}
}
\value{
m
}
\description{
Pick out a column from a flat matrix
}
