% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat-to-matrix.R
\name{flat_to_matrix}
\alias{flat_to_matrix}
\title{Turn flat matrix representation into a (dense) matrix}
\usage{
flat_to_matrix(x)
}
\arguments{
\item{x}{flat matrix representation from (e.g.) flatten_vector}
}
\value{
dense matrix matching x
}
\description{
Turn flat matrix representation into a (dense) matrix
}
