% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten-b-spline-covariate-repr.R
\name{flatten_b_spline_covariate_repr}
\alias{flatten_b_spline_covariate_repr}
\title{Generate a flat matrix from the spline representation}
\usage{
flatten_b_spline_covariate_repr(x)
}
\arguments{
\item{x}{spline representation}
}
\value{
flat matrix representation
}
\description{
Generate a flat matrix from the spline representation
}
