% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten-discrete-intercept-repr.R
\name{flatten_discrete_intercept_repr}
\alias{flatten_discrete_intercept_repr}
\title{Generate a flat matrix from the discrete intercept representation}
\usage{
flatten_discrete_intercept_repr(x)
}
\arguments{
\item{x}{discrete intercept representation}
}
\value{
flat matrix representation
}
\description{
Generate a flat matrix from the discrete intercept representation
}
