% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten-discrete-offset-repr.R
\name{flatten_discrete_offset_repr}
\alias{flatten_discrete_offset_repr}
\title{Generate a flat matrix from the discrete offset representation}
\usage{
flatten_discrete_offset_repr(x)
}
\arguments{
\item{x}{discrete offset representation}
}
\value{
flat matrix representation
}
\description{
Generate a flat matrix from the discrete offset representation
}
