% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten-linear-covariate-repr.R
\name{flatten_linear_covariate_repr}
\alias{flatten_linear_covariate_repr}
\title{List representation of linear covariate}
\usage{
flatten_linear_covariate_repr(x)
}
\arguments{
\item{x}{covariate representation}
}
\value{
flat version
}
\description{
List representation of linear covariate
}
