% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten-vector.R
\name{flatten_vector}
\alias{flatten_vector}
\title{Flat representation of a vector}
\usage{
flatten_vector(x)
}
\arguments{
\item{x}{vector to flatten}
}
\value{
row-wise flat vector representation
}
\description{
Flat representation of a vector
}
