% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knot-placer-basic-interval.R
\name{knot_placer_basic_interval}
\alias{knot_placer_basic_interval}
\title{Basic interval-spanning knot placement (not smart)}
\usage{
knot_placer_basic_interval(lb, ub, K = 4)
}
\arguments{
\item{lb}{lower bound of interval}

\item{ub}{upper bound of interval}

\item{K}{number of knots}
}
\value{
locations of all knots
}
\description{
If K is:
  Greater than 3: place knots at lower bound, 
    upper bound, outside knots one additional division 
    out. 
  Exactly 3: knots at lower bound, upper bound, and in 
    middle of interval
  Exactly 2: knots at lower and upper bound
  Exactly 1: knot at middle of interval
}
