% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-covariate.R
\name{linear_covariate}
\alias{linear_covariate}
\title{Simple continuous covariate}
\usage{
linear_covariate(x, label = "x")
}
\arguments{
\item{x}{value to use for covariate}

\item{label}{1-length character vector used as prefix for column name}
}
\value{
a matrix representing the covariate
}
\description{
Simple continuous covariate
}
