% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-covariate-repr.R
\name{linear_covariate_repr}
\alias{linear_covariate_repr}
\title{Simple continuous covariate representation}
\usage{
linear_covariate_repr(x, label = "x")
}
\arguments{
\item{x}{value to use for covariate}

\item{label}{1-length character vector used as prefix for column name}
}
\value{
a matrix representing the covariate
}
\description{
Simple continuous covariate representation
}
