% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition-all-reals.R,
%   R/partition-bounded-reals.R
\name{partition_all_reals}
\alias{partition_all_reals}
\title{Partition the real numbers into intervals}
\usage{
partition_all_reals(breaks = vector(mode = "numeric"))

partition_all_reals(breaks = vector(mode = "numeric"))
}
\arguments{
\item{breaks}{vector of breakpoints to create partitions with}
}
\value{
list of vectors, each vector contains the lower and upper
        bound in order

list of vectors, each vector contains the lower and upper
        bound in order
}
\description{
Partition the real numbers into intervals

Partition the real numbers into intervals
}
