% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition-bounded-reals.R
\name{partition_bounded_reals}
\alias{partition_bounded_reals}
\title{Partition a subset of the reals into intervals}
\usage{
partition_bounded_reals(breaks, lb, ub)
}
\arguments{
\item{breaks}{vector of breakpoints to create partitions with}

\item{lb}{lowest number to include}

\item{ub}{highest number to include}
}
\value{
list of vectors, each vector contains the lower and upper
        bound in order
}
\description{
Partition a subset of the reals into intervals
}
