% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten-matrix.R
\name{row_flatten_matrix}
\alias{row_flatten_matrix}
\title{Flat representation of a matrix, primarily for testing}
\usage{
row_flatten_matrix(x)
}
\arguments{
\item{x}{matrix to flatten}
}
\value{
row-wise flat matrix representation
}
\description{
Flat representation of a matrix, primarily for testing
}
