% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-column-to-constant.R
\name{set_column_to_constant}
\alias{set_column_to_constant}
\title{Label certain parameter (by name or index) as one that is *not* 
estimated. It's parameter will be fixed to 1.0}
\usage{
set_column_to_constant(m, name, idx)
}
\arguments{
\item{m}{model to update}

\item{name}{parameter name to update}

\item{idx}{index(es) of entries to update}
}
\description{
Label certain parameter (by name or index) as one that is *not* 
estimated. It's parameter will be fixed to 1.0
}
