% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-flat-linear-model-scaling.R
\name{set_flat_linear_model_scaling}
\alias{set_flat_linear_model_scaling}
\title{Label certain parameter (by name or index) with a constant scale}
\usage{
set_flat_linear_model_scaling(m, name, idx, scale)
}
\arguments{
\item{m}{model to update}

\item{name}{parameter name to update}

\item{idx}{index(es) of entries to update}

\item{scale}{scaling constant}
}
\description{
Label certain parameter (by name or index) with a constant scale
}
