% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-hierarchical-flat-linear-model-scaling.R
\name{set_hierarchical_flat_linear_model_scaling}
\alias{set_hierarchical_flat_linear_model_scaling}
\title{Label certain parameter (by name or index) with an estimated scale}
\usage{
set_hierarchical_flat_linear_model_scaling(m, name, idx)
}
\arguments{
\item{m}{model to update}

\item{name}{parameter name to update}

\item{idx}{index(es) of entries to update}
}
\description{
Label certain parameter (by name or index) with an estimated scale
}
