% Please edit documentation in R/datatables.R
\name{tableHeader}
\alias{tableHeader}
\title{Generate a table header or footer from column names}
\usage{
tableHeader(names, type = c("head", "foot"), escape = TRUE)
}
\arguments{
\item{names}{a character vector of the column names of the table (if it is an
object with column names, its column names will be used instead)}

\item{type}{generate a header (\samp{<thead></thead>}) or footer
(\samp{<tfoot></tfoot>})}

\item{escape}{whether to escape the names (see \code{\link{datatable}})}
}
\value{
A tag object generated by \code{htmltools::tags}.
}
\description{
A convenience function to generate a table header or footer given the column
names. It is basically a wrapper of \code{htmltools::tags$th} applied to the
column names.
}
\examples{
library(DT)
tableHeader(iris)  # or equivalently,
tableHeader(colnames(iris))
tableHeader(iris, "foot")  # footer

library(htmltools)
tags$table(tableHeader(iris, "head"), tableHeader(iris, "foot"))
}
